# app/api/automations/[automationId]/route.ts
import { NextResponse } from "next/server";
import { z } from "zod";

import { MAX_AUTOMATION_RUN_TIMEOUT_MINUTES } from "@/lib/constants";

import { requireUser } from "@/lib/auth";
import {
  assertFutureRunAt,
  deleteAutomation,
  getAutomation,
  updateAutomation
} from "@/lib/automations";
import { badRequest, ok, parseRouteParams } from "@/lib/http";
import { notifyConfigInputSchema } from "@/lib/notifications";
import { getPersona } from "@/lib/personas";
import { getProviderProfile } from "@/lib/settings";
import { getBot } from "@/lib/bots";

const paramsSchema = z.object({
  automationId: z.string().min(1)
});

const updateSchema = z.object({
  name: z.string().trim().min(1).max(100).optional(),
  prompt: z.string().trim().min(1).optional(),
  providerProfileId: z.string().min(1).optional(),
  personaId: z.string().min(1).nullable().optional(),
  botId: z.string().min(1).nullable().optional(),
  scheduleKind: z.enum(["interval", "calendar", "once"]).optional(),
  intervalMinutes: z.number().int().nullable().optional(),
  calendarFrequency: z.enum(["daily", "weekly"]).nullable().optional(),
  timeOfDay: z.string().nullable().optional(),
  daysOfWeek: z.array(z.number().int().min(0).max(6)).optional(),
  runAt: z.string().datetime({ offset: true }).nullable().optional(),
  continuePreviousConversation: z.boolean().optional(),
  enabled: z.boolean().optional(),
  research: z.boolean().optional(),
  runTimeoutMinutes: z.number().int().min(1).max(MAX_AUTOMATION_RUN_TIMEOUT_MINUTES).nullable().optional(),
  notifyConfig: notifyConfigInputSchema.optional()
}).refine(
  (value) => Object.keys(value).length > 0,
  "Invalid automation update"
);

export async function GET(
  _request: Request,
  context: { params: Promise<{ automationId: string }> }
) {
  const user = await requireUser();
    const params = await parseRouteParams(context, paramsSchema, "automation id");
  if (params instanceof NextResponse) return params;

  const automation = getAutomation(params.automationId, user.id);

  if (!automation) {
    return badRequest("Automation not found", 404);
  }

  return ok({ automation });
}

export async function PATCH(
  request: Request,
  context: { params: Promise<{ automationId: string }> }
) {
  const user = await requireUser();
    const params = await parseRouteParams(context, paramsSchema, "automation id");
  if (params instanceof NextResponse) return params;

  const body = updateSchema.safeParse(await request.json());

  if (!body.success) {
    return badRequest("Invalid automation update");
  }

  const automation = getAutomation(params.automationId, user.id);

  if (!automation) {
    return badRequest("Automation not found", 404);
  }

  if (body.data.providerProfileId && !getProviderProfile(body.data.providerProfileId)) {
    return badRequest("Provider profile not found", 404);
  }

  if (body.data.personaId && !getPersona(body.data.personaId, user.id)) {
    return badRequest("Persona not found", 404);
  }

  if (body.data.botId && !getBot(body.data.botId, user.id)) {
    return badRequest("Bot not found", 404);
  }

  try {
    if (body.data.runAt !== undefined || body.data.scheduleKind !== undefined) {
      assertFutureRunAt(
        body.data.scheduleKind ?? automation.scheduleKind,
        body.data.runAt === undefined ? automation.runAt : body.data.runAt
      );
    }
    const updated = updateAutomation(automation.id, body.data, user.id);
    return ok({ automation: updated });
  } catch (error) {
    return badRequest(error instanceof Error ? error.message : "Invalid automation update");
  }
}

export async function DELETE(
  _request: Request,
  context: { params: Promise<{ automationId: string }> }
) {
  const user = await requireUser();
    const params = await parseRouteParams(context, paramsSchema, "automation id");
  if (params instanceof NextResponse) return params;

  const deleted = deleteAutomation(params.automationId, user.id);
  if (!deleted) {
    return badRequest("Automation not found", 404);
  }

  return ok({ success: true });
}