# contracts/mobile-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eidon Mobile API",
    "version": "1.0.0",
    "description": "Platform-neutral native-client API. Dates are RFC 3339 strings and identifiers are opaque. This v1 contract is released together with the matching native client."
  },
  "servers": [{ "url": "/api/v1" }],
  "security": [{ "mobileBearer": [] }],
...
  "paths": {
...
    "/automations/{automationId}": {
      "parameters": [{ "$ref": "#/components/parameters/automationId" }],
      "get": { "operationId": "getAutomation", "tags": ["Automations"], "responses": { "200": { "$ref": "#/components/responses/Automation" } } },
      "patch": { "operationId": "updateAutomation", "tags": ["Automations"], "requestBody": { "$ref": "#/components/requestBodies/AutomationUpdate" }, "responses": { "200": { "$ref": "#/components/responses/Automation" } } },
      "delete": { "operationId": "deleteAutomation", "tags": ["Automations"], "responses": { "200": { "$ref": "#/components/responses/Success" } } }
    },
...
    "/messages/{messageId}/fork": { "post": { "operationId": "forkMessage", "tags": ["Messages"], "description": "Copies the conversation into a new one. Forking from an assistant reply keeps everything up to and including it; forking from a user message keeps everything before it and returns that message as the new conversation's composer draft. Bot home conversations cannot be forked.", "parameters": [{ "$ref": "#/components/parameters/messageId" }], "responses": { "201": { "$ref": "#/components/responses/ConversationFork" } } } },
...
    "/bots/{botId}": { "parameters": [{ "$ref": "#/components/parameters/botId" }], "get": { "operationId": "getBot", "tags": ["Agents"], "responses": { "200": { "$ref": "#/components/responses/BotDetail" } } }, "patch": { "operationId": "updateBot", "tags": ["Agents"], "requestBody": { "$ref": "#/components/requestBodies/BotUpdate" }, "responses": { "200": { "$ref": "#/components/responses/Bot" } } }, "delete": { "operationId": "deleteBot", "tags": ["Agents"], "responses": { "200": { "$ref": "#/components/responses/BotDeleted" } } } },
...
    "/bots/{botId}/skills/maintenance": { "parameters": [{ "$ref": "#/components/parameters/botId" }], "get": { "operationId": "getBotSkillMaintenance", "tags": ["Skills"], "description": "Reports the skill curator configuration, the curator's state for this bot's library, and the names of archived skills that can be restored.", "responses": { "200": { "$ref": "#/components/responses/BotSkillMaintenance" }, "404": { "$ref": "#/components/responses/Error" } } }, "post": { "operationId": "runBotSkillMaintenance", "tags": ["Skills"], "description": "Applies one library maintenance operation for this bot: pin or unpin a skill, adopt it as explicitly used, restore an archived skill, pause or resume the curator, or run a curation pass.", "requestBody": { "$ref": "#/components/requestBodies/BotSkillMaintenance" }, "responses": { "200": { "$ref": "#/components/responses/BotSkillMaintenanceResult" }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "securitySchemes": { "mobileBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "Eidon mobile session JWT" } },
    "parameters": {
...
      "automationId": { "name": "automationId", "in": "path", "required": true, "schema": { "type": "string" } },
      "botId": { "name": "botId", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "messageId": { "name": "messageId", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "requestBodies": {
...
      "AutomationUpdate": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AutomationUpdateRequest" } } } },
...
      "BotUpdate": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotUpdateRequest" } } } },
...
      "BotSkillMaintenance": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotSkillMaintenanceRequest" } } } },
...
    },
    "responses": {
...
      "Success": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessEnvelope" } } } },
      "Error": { "description": "Stable error response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "ConversationFork": { "description": "Newly created fork and the composer draft it should open with", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConversationForkEnvelope" } } } },
...
      "Automation": { "description": "Owned automation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AutomationEnvelope" } } } },
...
      "Bot": { "description": "Owned bot summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotEnvelope" } } } },
...
      "BotDetail": { "description": "Owned bot summary with recent runs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotDetailEnvelope" } } } },
      "BotDeleted": { "description": "Deleted bot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotDeletedEnvelope" } } } },
...
      "BotSkillMaintenance": { "description": "Skill curator state for a bot's library", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotSkillMaintenanceEnvelope" } } } },
      "BotSkillMaintenanceResult": { "description": "Outcome of one library maintenance operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BotSkillMaintenanceResultEnvelope" } } } },
...
    },
    "schemas": {
...
      "Rfc3339": { "type": "string", "format": "date-time" },
      "OpaqueId": { "type": "string", "minLength": 1, "maxLength": 512 },
      "NullableId": { "oneOf": [{ "$ref": "#/components/schemas/OpaqueId" }, { "type": "null" }] },
      "ReasoningEffort": { "type": "string", "enum": ["none", "low", "medium", "high", "xhigh", "max"] },
...
      "AutomationUpdateRequest": { "type": "object", "additionalProperties": false, "minProperties": 1, "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 100 }, "prompt": { "type": "string", "minLength": 1 }, "providerProfileId": { "$ref": "#/components/schemas/OpaqueId" }, "personaId": { "$ref": "#/components/schemas/NullableId" }, "botId": { "$ref": "#/components/schemas/NullableId" }, "scheduleKind": { "type": "string", "enum": ["interval", "calendar", "once"] }, "runAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "intervalMinutes": { "type": ["integer", "null"] }, "calendarFrequency": { "type": ["string", "null"], "enum": ["daily", "weekly", null] }, "timeOfDay": { "type": ["string", "null"] }, "daysOfWeek": { "type": "array", "uniqueItems": true, "items": { "type": "integer", "minimum": 0, "maximum": 6 } }, "continuePreviousConversation": { "type": "boolean" }, "enabled": { "type": "boolean" }, "research": { "type": "boolean" }, "runTimeoutMinutes": { "type": ["integer", "null"], "minimum": 1, "maximum": 720 }, "notifyConfig": { "$ref": "#/components/schemas/NotifyConfig" } } },
...
      "SuccessEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["success"], "properties": { "success": { "type": "boolean" } } } } },
      "ErrorEnvelope": { "type": "object", "additionalProperties": false, "required": ["error"], "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "details": { "type": ["object", "array", "string", "number", "boolean", "null"] } } } } },
...
      "Conversation": { "type": "object", "additionalProperties": false, "required": ["id", "title", "titleGenerationStatus", "folderId", "providerProfileId", "reasoningEffort", "automationId", "automationRunId", "conversationOrigin", "sortOrder", "createdAt", "updatedAt", "isActive", "shareEnabled", "sharedAt", "isTemporary"], "properties": { "id": { "$ref": "#/components/schemas/OpaqueId" }, "title": { "type": "string" }, "titleGenerationStatus": { "type": "string", "enum": ["pending", "running", "completed", "failed"] }, "folderId": { "$ref": "#/components/schemas/NullableId" }, "providerProfileId": { "$ref": "#/components/schemas/NullableId" }, "reasoningEffort": { "oneOf": [{ "$ref": "#/components/schemas/ReasoningEffort" }, { "type": "null" }] }, "automationId": { "$ref": "#/components/schemas/NullableId" }, "automationRunId": { "$ref": "#/components/schemas/NullableId" }, "conversationOrigin": { "type": "string", "enum": ["manual", "automation", "bot"] }, "sortOrder": { "type": "integer" }, "createdAt": { "$ref": "#/components/schemas/Rfc3339" }, "updatedAt": { "$ref": "#/components/schemas/Rfc3339" }, "isActive": { "type": "boolean" }, "shareEnabled": { "type": "boolean" }, "sharedAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "isTemporary": { "type": "boolean" }, "matchSnippet": { "type": "string" } } },
      "Attachment": { "type": "object", "additionalProperties": false, "required": ["id", "conversationId", "messageId", "filename", "mimeType", "byteSize", "sha256", "kind", "createdAt"], "properties": { "id": { "$ref": "#/components/schemas/OpaqueId" }, "conversationId": { "$ref": "#/components/schemas/OpaqueId" }, "messageId": { "$ref": "#/components/schemas/NullableId" }, "filename": { "type": "string" }, "mimeType": { "type": "string" }, "byteSize": { "type": "integer", "minimum": 0 }, "sha256": { "type": "string" }, "kind": { "type": "string", "enum": ["image", "text", "file"] }, "createdAt": { "$ref": "#/components/schemas/Rfc3339" } } },
...
      "Automation": { "type": "object", "additionalProperties": false, "required": ["id", "name", "prompt", "providerProfileId", "personaId", "scheduleKind", "intervalMinutes", "calendarFrequency", "timeOfDay", "daysOfWeek", "runAt", "continuePreviousConversation", "enabled", "research", "runTimeoutMinutes", "notifyConfig", "nextRunAt", "lastScheduledFor", "lastStartedAt", "lastFinishedAt", "lastStatus", "createdAt", "updatedAt"], "properties": { "id": { "$ref": "#/components/schemas/OpaqueId" }, "name": { "type": "string" }, "prompt": { "type": "string" }, "providerProfileId": { "$ref": "#/components/schemas/OpaqueId" }, "personaId": { "$ref": "#/components/schemas/NullableId" }, "botId": { "$ref": "#/components/schemas/NullableId" }, "scheduleKind": { "type": "string", "enum": ["interval", "calendar", "once"] }, "intervalMinutes": { "type": ["integer", "null"] }, "calendarFrequency": { "type": ["string", "null"], "enum": ["daily", "weekly", null] }, "timeOfDay": { "type": ["string", "null"] }, "daysOfWeek": { "type": "array", "items": { "type": "integer", "minimum": 0, "maximum": 6 } }, "runAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "continuePreviousConversation": { "type": "boolean" }, "enabled": { "type": "boolean" }, "research": { "type": "boolean" }, "runTimeoutMinutes": { "type": ["integer", "null"], "minimum": 1, "maximum": 720 }, "nextRunAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "lastScheduledFor": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "lastStartedAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "lastFinishedAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "lastStatus": { "type": ["string", "null"], "enum": ["queued", "running", "completed", "failed", "missed", "stopped", "paused", null] }, "createdAt": { "$ref": "#/components/schemas/Rfc3339" }, "updatedAt": { "$ref": "#/components/schemas/Rfc3339" }, "notifyConfig": { "$ref": "#/components/schemas/NotifyConfig" } } },
      "NotifySecretRef": { "type": "object", "additionalProperties": false, "required": ["set"], "properties": { "set": { "type": "boolean", "const": true } } },
      "NtfyNotifyChannel": { "type": "object", "additionalProperties": false, "required": ["kind", "topic"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "const": "ntfy" }, "server": { "type": "string" }, "topic": { "type": "string", "minLength": 1, "maxLength": 128 }, "priority": { "type": "integer", "minimum": 1, "maximum": 5 }, "includeSummary": { "type": "boolean" } } },
      "WebhookNotifyChannel": { "type": "object", "additionalProperties": false, "required": ["kind", "url"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "const": "webhook" }, "url": { "oneOf": [{ "type": "string", "minLength": 1 }, { "$ref": "#/components/schemas/NotifySecretRef" }] }, "headers": { "type": "object", "additionalProperties": { "oneOf": [{ "type": "string" }, { "$ref": "#/components/schemas/NotifySecretRef" }] } }, "includeSummary": { "type": "boolean" } } },
      "PushoverNotifyChannel": { "type": "object", "additionalProperties": false, "required": ["kind"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "const": "pushover" }, "device": { "type": "string", "minLength": 1, "maxLength": 64 }, "priority": { "type": "integer", "minimum": -2, "maximum": 1 }, "includeSummary": { "type": "boolean" } } },
      "PushNotifyChannel": { "type": "object", "additionalProperties": false, "required": ["kind"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "const": "push" }, "includeSummary": { "type": "boolean" } } },
      "NotifyConfig": { "type": "object", "additionalProperties": false, "required": ["channels"], "properties": { "channels": { "type": "array", "maxItems": 10, "items": { "oneOf": [{ "$ref": "#/components/schemas/NtfyNotifyChannel" }, { "$ref": "#/components/schemas/WebhookNotifyChannel" }, { "$ref": "#/components/schemas/PushoverNotifyChannel" }, { "$ref": "#/components/schemas/PushNotifyChannel" }] } } } },
...
      "BotSummary": { "type": "object", "additionalProperties": false, "required": ["id", "name", "title", "description", "avatarSeed", "isChief", "homeConversationId", "providerProfileId", "status", "waitingForInput", "unread", "lastRunAt", "createdAt", "updatedAt"], "properties": { "id": { "$ref": "#/components/schemas/OpaqueId" }, "name": { "type": "string" }, "title": { "type": "string" }, "description": { "type": "string" }, "avatarSeed": { "type": "string" }, "isChief": { "type": "boolean" }, "homeConversationId": { "$ref": "#/components/schemas/OpaqueId" }, "providerProfileId": { "$ref": "#/components/schemas/NullableId" }, "status": { "type": "string", "enum": ["idle", "queued", "running", "waiting_user"] }, "waitingForInput": { "type": "boolean" }, "unread": { "type": "boolean" }, "lastRunAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "createdAt": { "$ref": "#/components/schemas/Rfc3339" }, "updatedAt": { "$ref": "#/components/schemas/Rfc3339" } } },
      "BotRun": { "type": "object", "additionalProperties": false, "required": ["id", "botId", "conversationId", "triggerSource", "status", "startedAt", "finishedAt", "parentMessageId", "requestedByBotId", "errorMessage", "createdAt"], "properties": { "id": { "$ref": "#/components/schemas/OpaqueId" }, "botId": { "$ref": "#/components/schemas/OpaqueId" }, "conversationId": { "$ref": "#/components/schemas/OpaqueId" }, "triggerSource": { "type": "string", "enum": ["dm", "delegated", "routine"] }, "status": { "type": "string", "enum": ["queued", "running", "waiting_user", "completed", "failed", "stopped"] }, "startedAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "finishedAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "parentMessageId": { "$ref": "#/components/schemas/NullableId" }, "requestedByBotId": { "$ref": "#/components/schemas/NullableId" }, "errorMessage": { "type": ["string", "null"] }, "createdAt": { "$ref": "#/components/schemas/Rfc3339" } } },
...
      "BotUpdateRequest": { "type": "object", "additionalProperties": false, "minProperties": 1, "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 60 }, "title": { "type": "string", "maxLength": 120 }, "description": { "type": "string", "maxLength": 1000 }, "systemPrompt": { "type": "string", "maxLength": 64000 }, "providerProfileId": { "$ref": "#/components/schemas/NullableId" } } },
...
      "ComposerDraft": { "type": "object", "additionalProperties": false, "required": ["content", "attachments"], "properties": { "content": { "type": "string" }, "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } } } },
      "NullableComposerDraft": { "oneOf": [{ "$ref": "#/components/schemas/ComposerDraft" }, { "type": "null" }] },
      "ConversationForkEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["conversation", "draft"], "properties": { "conversation": { "$ref": "#/components/schemas/Conversation" }, "draft": { "$ref": "#/components/schemas/NullableComposerDraft" } } } } },
...
      "AutomationEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["automation"], "properties": { "automation": { "$ref": "#/components/schemas/Automation" } } } } },
...
      "BotEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["bot"], "properties": { "bot": { "$ref": "#/components/schemas/BotSummary" } } } } },
...
      "BotDetailEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["bot", "runs"], "properties": { "bot": { "$ref": "#/components/schemas/BotSummary" }, "runs": { "type": "array", "items": { "$ref": "#/components/schemas/BotRun" } } } } } },
      "BotDeletedEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["deleted"], "properties": { "deleted": { "type": "boolean" } } } } },
...
      "SkillCuratorBackupConfig": { "type": "object", "additionalProperties": false, "required": ["enabled", "keep"], "properties": { "enabled": { "type": "boolean" }, "keep": { "type": "integer", "minimum": 0 } } },
      "SkillCuratorBackgroundReviewConfig": { "type": "object", "additionalProperties": false, "required": ["enabled"], "properties": { "enabled": { "type": "boolean" } } },
      "SkillCuratorConfig": { "type": "object", "additionalProperties": false, "required": ["enabled", "intervalHours", "minIdleHours", "staleAfterDays", "archiveAfterDays", "consolidate", "pruneBuiltins", "archiveTtlDays", "backup", "backgroundReview", "ledger"], "properties": { "enabled": { "type": "boolean" }, "intervalHours": { "type": "integer", "minimum": 1 }, "minIdleHours": { "type": "number", "minimum": 0 }, "staleAfterDays": { "type": "integer", "minimum": 1 }, "archiveAfterDays": { "type": "integer", "minimum": 1 }, "consolidate": { "type": "boolean" }, "pruneBuiltins": { "type": "boolean" }, "archiveTtlDays": { "type": "integer", "minimum": 0, "description": "Zero keeps archived skills indefinitely." }, "backup": { "$ref": "#/components/schemas/SkillCuratorBackupConfig" }, "backgroundReview": { "$ref": "#/components/schemas/SkillCuratorBackgroundReviewConfig" }, "ledger": { "type": "boolean" } } },
...
      "CuratorState": { "type": "object", "additionalProperties": false, "required": ["seededAt", "lastRunAt", "lastInteractionAt", "paused", "lastSummary"], "properties": { "seededAt": { "$ref": "#/components/schemas/Rfc3339" }, "lastRunAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "lastInteractionAt": { "oneOf": [{ "$ref": "#/components/schemas/Rfc3339" }, { "type": "null" }] }, "paused": { "type": "boolean" }, "lastSummary": { "type": ["string", "null"] } } },
      "BotSkillMaintenanceEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["config", "state", "archived"], "properties": { "config": { "$ref": "#/components/schemas/SkillCuratorConfig" }, "state": { "$ref": "#/components/schemas/CuratorState" }, "archived": { "type": "array", "items": { "type": "string" } } } } } },
      "BotSkillMaintenanceRequest": { "oneOf": [
        { "type": "object", "additionalProperties": false, "required": ["op", "ref"], "properties": { "op": { "const": "pin" }, "ref": { "$ref": "#/components/schemas/OpaqueId" } } },
        { "type": "object", "additionalProperties": false, "required": ["op", "ref"], "properties": { "op": { "const": "unpin" }, "ref": { "$ref": "#/components/schemas/OpaqueId" } } },
        { "type": "object", "additionalProperties": false, "required": ["op", "ref"], "properties": { "op": { "const": "adopt" }, "ref": { "$ref": "#/components/schemas/OpaqueId" } } },
        { "type": "object", "additionalProperties": false, "required": ["op", "archivedName"], "properties": { "op": { "const": "restore" }, "archivedName": { "type": "string", "minLength": 1 } } },
        { "type": "object", "additionalProperties": false, "required": ["op"], "properties": { "op": { "const": "pause" } } },
        { "type": "object", "additionalProperties": false, "required": ["op"], "properties": { "op": { "const": "resume" } } },
        { "type": "object", "additionalProperties": false, "required": ["op"], "properties": { "op": { "const": "run" }, "dryRun": { "type": "boolean" } } }
      ] },
      "BotSkillMaintenanceResultEnvelope": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "oneOf": [
        { "type": "object", "additionalProperties": false, "required": ["ref", "pinned"], "properties": { "ref": { "$ref": "#/components/schemas/OpaqueId" }, "pinned": { "type": "boolean" } } },
        { "type": "object", "additionalProperties": false, "required": ["ref", "adopted"], "properties": { "ref": { "$ref": "#/components/schemas/OpaqueId" }, "adopted": { "type": "boolean" } } },
        { "type": "object", "additionalProperties": false, "required": ["ref", "archived"], "properties": { "ref": { "$ref": "#/components/schemas/OpaqueId" }, "archived": { "type": "array", "items": { "type": "string" } } } },
        { "type": "object", "additionalProperties": false, "required": ["paused"], "properties": { "paused": { "type": "boolean" } } },
        { "type": "object", "additionalProperties": false, "required": ["ran", "summary", "config", "archived"], "properties": { "ran": { "type": "boolean" }, "summary": { "type": ["string", "null"] }, "reason": { "type": "string" }, "config": { "$ref": "#/components/schemas/SkillCuratorConfig" }, "archived": { "type": "array", "items": { "type": "string" } } } }
      ] } } },
...
    }
  }
}